requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
autopep8==2.0.1
billiard==3.6.4.0
black==23.3.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==4.1.4
django-js-asset==2.0.0
django-mptt==0.14.0
djangorestframework==3.14.0
factory-boy==3.2.1
Faker==16.6.1
isort==5.12.0
kombu==5.2.4
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.1
Pillow==9.3.0
platformdirs==3.2.0
prompt-toolkit==3.0.36
psycopg2-binary==2.9.5
pycodestyle==2.10.0
python-dateutil==2.8.2
pytz==2022.6
redis==4.4.0
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
vine==5.0.0
wcwidth==0.2.5
